+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Experimental creation of [Zod schemas](https://zod.dev/)
+- Automatic generation of API clients
 
 ### Changed